pyproject.toml
[tool.poetry]
name = "library"
version = "1.0"
description = ""
authors = ["Alexandre Silva <alexandresilva933@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "^3.5.0"
attrs = "^21.4.0"
certifi = "^2021.10.8"
charset-normalizer = "^2.0.12"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
idna = "^3.3"
inflection = "^0.5.1"
iniconfig = "^1.1.1"
itypes = "1.2.0"
Jinja2 = "^3.0.3"
MarkupSafe = "^2.1.1"
packaging = "^21.3"
pluggy = "^1.0.0"
psycopg2-binary = "^2.9.3"
py = "^1.11.0"
pyparsing = "^3.0.7"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
pytz = "^2022.1"
requests = "^2.27.1"
ruamel-yaml = "^0.17.21"
ruamel-yaml-clib = "^0.2.6"
six = "^1.16.0"
sqlparse = "^0.4.2"
tomli = "^2.0.1"
uritemplate = "^4.1.1"
urllib3 = "1.26.9"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "library.settings.dev"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


